+++ hubdle-gradle-plugin/main/kotlin/com/javiersc/hubdle/project/extensions/config/publishing/maven/HubdleConfigPublishingMavenExtension.kt
@@ -15,7 +15,6 @@ import com.javiersc.hubdle.project.extensions.config.publishing.hubdlePublishing
 import com.javiersc.hubdle.project.extensions.config.publishing.isSemver
 import com.javiersc.hubdle.project.extensions.config.versioning.semver._internal.isTagPrefixProject
 import com.javiersc.hubdle.project.extensions.shared.features.gradle.hubdleGradlePluginFeature
-import com.javiersc.hubdle.project.extensions.shared.features.gradle.hubdleGradleVersionCatalog
 import javax.inject.Inject
 import org.gradle.api.Action
 import org.gradle.api.Project
@@ -89,9 +88,7 @@ internal fun HubdleConfigurableExtension.configurableMavenPublishing(
             }
         }
         val isNotGradlePlugin: Boolean = hubdleGradlePluginFeature.isFullEnabled.orNull == false
-        val isNotGradleVersionCatalog: Boolean =
-            hubdleGradleVersionCatalog.isFullEnabled.orNull == false
-        if (isNotGradlePlugin && isNotGradleVersionCatalog && mavenPublicationName != null) {
+        if (isNotGradlePlugin && mavenPublicationName != null) {
             configureMavenPublication(mavenPublicationName)
         }
         if (configEmptyJavaDocs) {